+++ swing/common-model/src/main/java/is/codion/swing/common/model/component/table/DefaultFilterTableSelectionModel.java
@@ -93,7 +93,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 
 	@Override
 	public void selectAll() {
-		setSelectionInterval(0, items.visible().size() - 1);
+		setSelectionInterval(0, items.visible().count() - 1);
 	}
 
 	@Override
@@ -270,7 +270,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 		@Override
 		public void set(Integer index) {
 			requireNonNull(index);
-			checkIndex(index, items.visible().size());
+			checkIndex(index, items.visible().count());
 			setSelectionInterval(index, index);
 		}
 
@@ -317,20 +317,20 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 
 		@Override
 		public void add(int index) {
-			checkIndex(index, items.visible().size());
+			checkIndex(index, items.visible().count());
 			addSelectionInterval(index, index);
 		}
 
 		@Override
 		public void remove(int index) {
-			checkIndex(index, items.visible().size());
+			checkIndex(index, items.visible().count());
 			removeSelectionInterval(index, index);
 		}
 
 		@Override
 		public void remove(Collection<Integer> indexes) {
 			indexes.forEach(index -> {
-				checkIndex(index, items.visible().size());
+				checkIndex(index, items.visible().count());
 				removeSelectionInterval(index, index);
 			});
 		}
@@ -355,7 +355,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 
 		@Override
 		public void moveUp() {
-			int visibleSize = items.visible().size();
+			int visibleSize = items.visible().count();
 			if (visibleSize > 0) {
 				int lastIndex = visibleSize - 1;
 				if (isSelectionEmpty()) {
@@ -371,7 +371,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 
 		@Override
 		public void moveDown() {
-			int filteredSize = items.visible().size();
+			int filteredSize = items.visible().count();
 			if (filteredSize > 0) {
 				if (isSelectionEmpty()) {
 					setSelectionInterval(0, 0);
@@ -390,7 +390,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 		}
 
 		private void checkIndexes(Collection<Integer> indexes) {
-			int size = items.visible().size();
+			int size = items.visible().count();
 			for (Integer index : indexes) {
 				checkIndex(index, size);
 			}
@@ -408,7 +408,7 @@ final class DefaultFilterTableSelectionModel<R> implements FilterTableSelectionM
 		@Override
 		public R get() {
 			int index = selectedIndex.get();
-			if (index >= 0 && index < items.visible().size()) {
+			if (index >= 0 && index < items.visible().count()) {
 				return items.visible().itemAt(index);
 			}