+++ swing/common-ui/src/main/java/is/codion/swing/common/ui/component/table/DefaultFilterTableSearchModel.java
@@ -171,12 +171,12 @@ final class DefaultFilterTableSearchModel<C> implements FilterTableSearchModel {
 
 	private void performSearch() {
 		clearSearchResults();
-		if (searchPredicate.isNull() || tableModel.items().visible().size() == 0 || tableModel.getColumnCount() == 0) {
+		if (searchPredicate.isNull() || tableModel.items().visible().count() == 0 || tableModel.getColumnCount() == 0) {
 			return;
 		}
 		Predicate<String> predicate = searchPredicate.get();
 		List<FilterTableColumn<C>> visibleColumns = columnModel.visible();
-		for (int row = 0; row < tableModel.items().visible().size(); row++) {
+		for (int row = 0; row < tableModel.items().visible().count(); row++) {
 			for (int columnIndex = 0; columnIndex < visibleColumns.size(); columnIndex++) {
 				FilterTableColumn<C> column = visibleColumns.get(columnIndex);
 				if (predicate.test(tableModel.getStringAt(row, column.identifier()))) {