+++ swing/framework-model/src/test/java/is/codion/swing/framework/model/SwingEntityTableModelTest.java
@@ -94,7 +94,7 @@ public final class SwingEntityTableModelTest extends AbstractEntityTableModelTes
 						testModel.filterModel().conditionModel(Detail.STRING);
 		filterModel.operands().equal().set("a");
 		testModel.items().filter();
-		assertEquals(4, testModel.items().filtered().size());
+		assertEquals(4, testModel.items().filtered().count());
 	}
 
 	@Test